+++ CHANGELOG.rst
@@ -11,9 +11,15 @@ New Features
 ^^^^^^^^^^^^
 - Add statistics for SAR imaging + fix variance of GammaNoise in doc (:gh:`740` by `Louise Friot Giroux`)
 
+- dinv.io functions for loading DICOM, NIFTI, COS, GEOTIFF etc. (:gh:`768` by `Andrew Wang`_)
+
 Changed
 ^^^^^^^
+- load_np_url now returns tensors
+- load_url helper function moved to io
 - (Breaking) Make HDF5Dataset similar to Trainer in the unsupervised setting by using NaNs for ground truths instead of a copy of the measurements (:gh:`761` by `Jérémy Scanvic`_)
+- utils/signal.py renamed to signals.py to avoid stdlib conflict (:gh:`768` by `Andrew Wang`_)
+- utils.get_data_home now creates folder if not exist (:gh:`768` by `Andrew Wang`_)
 
 Fixed
 ^^^^^